+++ adventure/src/test/java/com/github/idankoblik/SongQueueTest.java
@@ -3,6 +3,7 @@ package com.github.idankoblik;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 public class SongQueueTest extends AbstractSpigotSongTest {
@@ -116,8 +117,8 @@ public class SongQueueTest extends AbstractSpigotSongTest {
 
     @Override
     protected void initializeSongs() throws Exception {
-        this.song = new Song(this.nbsSong, DEFAULT_NOTE, audience(player));
-        this.locationSong = new Song(this.nbsSong, DEFAULT_NOTE, musicPosition, audience(world));
+        this.song = new Song(this.nbsSong, DEFAULT_NOTE, audience(player), null);
+        this.locationSong = new Song(this.nbsSong, DEFAULT_NOTE, audience(world), musicPosition);
 
         this.locationSongQueue = new SongQueue(DEFAULT_NOTE, musicPosition, audience(player));
         this.songQueue = new SongQueue(DEFAULT_NOTE, audience(player));
