+++ settings.gradle.kts
@@ -1,3 +1,4 @@
 rootProject.name = "jukebox"
 include("shared")
 include("adventure")
+include("discord")
